+++ tests/test_events.py
@@ -2,6 +2,7 @@ from __future__ import unicode_literals
 
 import mock
 
+from mopidy.core.playback import PlaybackState
 from mopidy.models import Playlist, TlTrack
 
 import pytest
@@ -61,6 +62,18 @@ def test_track_playback_ended_changes_playback_status_and_metadata(frontend):
 
     frontend.track_playback_ended(
         tl_track=TlTrack(), time_position=0)
+
+    frontend.mpris.player.PropertiesChanged.assert_called_with(
+        player.Player.INTERFACE,
+        {'Metadata': '...', 'PlaybackStatus': 'Stopped'}, [])
+
+
+def test_playback_state_changed_changes_playback_status_and_metadata(frontend):
+    frontend.mpris.player.Metadata = '...'
+    frontend.mpris.player.PlaybackStatus = 'Stopped'
+
+    frontend.playback_state_changed(
+        PlaybackState.PLAYING, PlaybackState.STOPPED)
 
     frontend.mpris.player.PropertiesChanged.assert_called_with(
         player.Player.INTERFACE,